Change sign-up to send strings, pull officer rating and comments.
Citizen sign-up sends strings for spinner fields rather than integers. When officer
checks his ratings, they are pulled from firebase. This is currently hardcoded to a
specific officer in a specific department, because log-in on officer is hardcoded
to a specific account. Comments-pulling is having issues showing the array that
is updatd with the new comments.