Citizen can upload insurance to Firebase and can zoom in on insurance
In initial contact screen, citizen can use the send documents button to send
their insurance document to firebase storage. Needs to update the beacon in
the database with the location of the uploaded insurance. When the citizen
uploads his insurance image on the app, he can now zoom in on the image.
location of the insurance.