Let citizen send all documents on initial contact screen.
Citizen can send all 3 documents if they are all uploaded to Firebase
Storage. Each document is saved in a separate folder in Storage,
with the name of the file being the same as the driver's uid. Documents
take a little while to upload, so need to include a progress spinner.