Build history tab for citizens.
Citizens can view all stops in a table and can click on a stop
to see details of the stop. The details are the location on a map, the
date and time of the stop, the address of the stop, the name of the officer
who stopped him, an image of the officer, and a reason for stopping.